requirements.txt
scrapy
pymongo
baidu-aip
pillow
requests